requirements.txt
pyserial
numpy
sounddevice
soundfile